Editor Refactor Followup
