+++ src/components/Inputs/RichTextEditor/RichTextEditor.stories.tsx
@@ -12,7 +12,7 @@ import {
   RichTextEditorFeatures,
 } from './RichTextEditor.types';
 import { amplify_h2, amplify_h3 } from 'src/components/Icons/AmplifyIcons';
-import { useFeatures } from 'src/hooks/useFeatures';
+import useFeatures from 'src/utils/getFeatures';
 
 export default {
   title: 'Inputs/RichTextEditor',
