+++ src/hooks/index.ts
@@ -16,7 +16,7 @@ import { usePrevious } from './usePrevious';
 import { useReleaseNotesQuery } from './useReleaseNotesQuery';
 import { useSignalRMessages } from './useSignalRMessages';
 import { useAmplifyKit } from 'src/hooks/useAmplifyKit';
-import { useFeatures } from 'src/hooks/useFeatures';
+import useFeatures from 'src/utils/getFeatures';
 import { useSelect } from 'src/hooks/useSelect';
 
 export {
