+++ src/hooks/useAmplifyKit.ts
@@ -11,6 +11,7 @@ interface AmplifyKitProps {
   placeholder?: string;
   onImageUpload?: OnImageUploadFn;
 }
+
 /* c8 ignore start */
 export const useAmplifyKit = ({
   features,
